+++ backend/src/main/java/reviewme/review/controller/ReviewApi.java
@@ -16,8 +16,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
-import reviewme.review.dto.request.CreateReviewRequest;
-import reviewme.review.dto.response.ReceivedReviewsResponse;
+import reviewme.review.service.dto.request.CreateReviewRequest;
+import reviewme.review.service.dto.response.list.ReceivedReviewsResponse;
 import reviewme.review.service.dto.response.detail.TemplateAnswerResponse;
 
 @Tag(name = "리뷰 관리")
