+++ backend/src/main/java/reviewme/review/controller/ReviewController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import reviewme.global.HeaderProperty;
-import reviewme.review.dto.request.CreateReviewRequest;
-import reviewme.review.dto.response.ReceivedReviewsResponse;
+import reviewme.review.service.dto.request.CreateReviewRequest;
+import reviewme.review.service.dto.response.list.ReceivedReviewsResponse;
 import reviewme.review.service.CreateReviewService;
 import reviewme.review.service.ReviewDetailLookupService;
 import reviewme.review.service.ReviewService;
