+++ backend/src/main/java/reviewme/review/service/CreateReviewService.java
@@ -13,8 +13,8 @@ import reviewme.review.domain.CheckboxAnswer;
 import reviewme.review.domain.Review;
 import reviewme.review.domain.TextAnswer;
 import reviewme.review.domain.exception.ReviewGroupNotFoundByRequestReviewCodeException;
-import reviewme.review.dto.request.CreateReviewAnswerRequest;
-import reviewme.review.dto.request.CreateReviewRequest;
+import reviewme.review.service.dto.request.CreateReviewAnswerRequest;
+import reviewme.review.service.dto.request.CreateReviewRequest;
 import reviewme.question.repository.QuestionRepository;
 import reviewme.review.repository.ReviewRepository;
 import reviewme.review.service.exception.SubmittedQuestionAndProvidedQuestionMismatchException;
