+++ backend/src/main/java/reviewme/review/service/ReviewService.java
@@ -11,9 +11,9 @@ import reviewme.review.domain.CheckboxAnswer;
 import reviewme.review.domain.Review;
 import reviewme.review.domain.exception.CategoryOptionByReviewNotFoundException;
 import reviewme.review.domain.exception.ReviewGroupNotFoundByGroupAccessCodeException;
-import reviewme.review.dto.response.ReceivedReviewCategoryResponse;
-import reviewme.review.dto.response.ReceivedReviewResponse;
-import reviewme.review.dto.response.ReceivedReviewsResponse;
+import reviewme.review.service.dto.response.list.ReceivedReviewCategoryResponse;
+import reviewme.review.service.dto.response.list.ReceivedReviewResponse;
+import reviewme.review.service.dto.response.list.ReceivedReviewsResponse;
 import reviewme.question.repository.QuestionRepository;
 import reviewme.review.repository.ReviewRepository;
 import reviewme.reviewgroup.domain.ReviewGroup;
