+++ backend/src/main/java/reviewme/reviewgroup/service/ReviewGroupService.java
@@ -4,8 +4,8 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import reviewme.reviewgroup.domain.ReviewGroup;
-import reviewme.reviewgroup.dto.ReviewGroupCreationRequest;
-import reviewme.reviewgroup.dto.ReviewGroupCreationResponse;
+import reviewme.reviewgroup.service.dto.ReviewGroupCreationRequest;
+import reviewme.reviewgroup.service.dto.ReviewGroupCreationResponse;
 import reviewme.reviewgroup.repository.ReviewGroupRepository;
 
 @Service
