+++ backend/src/main/java/reviewme/reviewgroup/service/dto/ReviewGroupCreationRequest.java
@@ -1,4 +1,4 @@
-package reviewme.reviewgroup.dto;
+package reviewme.reviewgroup.service.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
