+++ backend/src/main/java/reviewme/template/controller/TemplateController.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import reviewme.template.dto.response.TemplateResponse;
+import reviewme.template.service.dto.response.TemplateResponse;
 import reviewme.template.service.TemplateService;
 
 @RestController
