+++ backend/src/main/java/reviewme/template/service/dto/response/SectionResponse.java
@@ -1,4 +1,4 @@
-package reviewme.template.dto.response;
+package reviewme.template.service.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.annotation.Nullable;
