+++ backend/src/test/java/reviewme/review/service/CreateReviewServiceTest.java
@@ -13,8 +13,8 @@ import reviewme.question.domain.Question;
 import reviewme.question.domain.QuestionType;
 import reviewme.question.repository.OptionGroupRepository;
 import reviewme.question.repository.OptionItemRepository;
-import reviewme.review.dto.request.CreateReviewAnswerRequest;
-import reviewme.review.dto.request.CreateReviewRequest;
+import reviewme.review.service.dto.request.CreateReviewAnswerRequest;
+import reviewme.review.service.dto.request.CreateReviewRequest;
 import reviewme.review.repository.CheckboxAnswerRepository;
 import reviewme.question.repository.QuestionRepository;
 import reviewme.review.repository.ReviewRepository;
