+++ backend/src/test/java/reviewme/review/service/ReviewServiceTest.java
@@ -16,8 +16,8 @@ import reviewme.question.repository.OptionItemRepository;
 import reviewme.review.domain.CheckboxAnswer;
 import reviewme.review.domain.Review;
 import reviewme.review.domain.exception.ReviewGroupNotFoundByGroupAccessCodeException;
-import reviewme.review.dto.response.ReceivedReviewCategoryResponse;
-import reviewme.review.dto.response.ReceivedReviewsResponse;
+import reviewme.review.service.dto.response.list.ReceivedReviewCategoryResponse;
+import reviewme.review.service.dto.response.list.ReceivedReviewsResponse;
 import reviewme.review.repository.CheckboxAnswerRepository;
 import reviewme.question.repository.QuestionRepository;
 import reviewme.review.repository.ReviewRepository;
